+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Sort tasks in this order: 1. priority, 2. time, 3. creation date (newest first)
 * Remove duplicate tasks from the daily note
 * Sync completed tasks from Obsidian to Todoist
+* Move amount of tasks under each heading
 
 ### 1.0.5: 2025-01-16